+++ test/CountReadableStream.mjs
@@ -1,6 +1,4 @@
-'use strict';
-
-const { Readable } = require('stream');
+import { Readable } from 'stream';
 
 /**
  * A count readable stream, for testing purposes.
@@ -9,7 +7,7 @@ const { Readable } = require('stream');
  * @see [Example counting stream in the Node.js docs](https://nodejs.org/api/stream.html#stream_an_example_counting_stream).
  * @ignore
  */
-module.exports = class CountReadableStream extends Readable {
+export default class CountReadableStream extends Readable {
   constructor(options) {
     super(options);
     this._max = 1000000;
@@ -20,4 +18,4 @@ module.exports = class CountReadableStream extends Readable {
     const i = this._index++;
     this.push(i > this._max ? null : Buffer.from(String(i), 'ascii'));
   }
-};
+}