+++ test/listen.mjs
@@ -1,5 +1,3 @@
-'use strict';
-
 /**
  * Starts a Node.js HTTP server.
  * @kind function
@@ -8,7 +6,7 @@
  * @returns {Promise<{port: number, close: Function}>} Resolves the port the server is listening on, and a server close function.
  * @ignore
  */
-module.exports = function listen(server) {
+export default function listen(server) {
   return new Promise((resolve, reject) => {
     server.listen(function (error) {
       if (error) reject(error);
@@ -19,4 +17,4 @@ module.exports = function listen(server) {
         });
     });
   });
-};
+}