+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Updated Node.js support to `^12.20 || >= 14.13`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
+- The tests are now ESM in `.mjs` files instead of CJS in `.js` files.
 
 ### Patch